requirements.txt
asgiref==3.2.3
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.0.3
djangorestframework==3.11.0
gunicorn==20.0.4
mysql-connector-python==8.0.17
mysqlclient==1.4.6
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
